going through the list
