+++ include/Channel.hpp
@@ -5,39 +5,54 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: mmakagon <mmakagon@student.42.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/11/05 14:48:02 by pyerima           #+#    #+#             */
-/*   Updated: 2024/11/18 18:40:41 by mmakagon         ###   ########.fr       */
+/*   Created: 2024/11/04 14:58:10 by mmakagon          #+#    #+#             */
+/*   Updated: 2024/11/19 02:28:53 by mmakagon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef CHANNEL_HPP
-# define CHANNEL_HPP
+#define CHANNEL_HPP
 
 #include <iostream>
 #include <set>
-#include <arpa/inet.h>
+#include "Client.hpp"
+#include "ft_irc.hpp"
 
 class Channel {
+	private:
+		std::string					name;
+		std::string					topic;
+		unsigned int				hashed_pass;
+		std::set<const Client*>		clients;
+		std::set<const Client*>		admins;
+		size_t						clnts_limit;
+		bool						invite_only;
+		bool						topic_admins_only;
+
+		ssize_t		singleMessage(const Client& client, const std::string message) const;
+
 	public:
-		Channel(const std::string& name);
-		//Channel ( const Channel &other );
-		~Channel( void );
+		Channel(const std::string& in_name, const Client& creator);
+		~Channel();
 
-		std::string name;
-		std::set<int> clients; //client file descriptors
-		std::string topic;
-		std::string password;
-		bool invite_only;
-		bool topic_restricted;
-		int user_limit;
+		const std::string&	getName(void) const;
+		const std::string&	getTopic(void) const;
 
-		void join(int client_fd);
-		void part(int client_fd);
-		void sendMessage(const std::string& message, int sender_fd);
-		void setTopic(const std::string& new_topic);
-		std::string getTopic() const;
-		void setPassword(const std::string& pass);
-		bool isOperator(int client_fd);
+
+		// bool here and in other functions to check if an action was successfull, could be changed to exceptions
+		bool		setTopic(const std::string& in_topic, const Client& admin);
+		bool		setPass(std::string& in_pass, const Client& admin);
+
+		bool		addClient(const Client& in_client, const Client& admin);
+		bool		addAdmin(const Client& in_client, const Client& admin);
+
+		bool		kickClient(const Client& in_client, const Client& admin);
+		bool		kickAdmin(const Client& in_client, const Client& admin);
+
+		bool		joinChannel(const Client& in_client);
+		bool		leaveChannel(const Client& in_client);
+
+		bool		changeMode(const char* const in_mode, const Client& admin);
 };
 
 #endif
